+++ src/main/java/com/fastcampus/jcjboard/paging/Paging.java
@@ -1,7 +1,7 @@
 package com.fastcampus.jcjboard.paging;
 
 public class Paging {
-    private PerPage perPage; // page, perPageNum 을 가지고 있음
+//    private PerPage perPage; // page, perPageNum 을 가지고 있음
 
     private int totalCount; // 전체 게시글 수
     private int startPage; // 게시글 번호에 따른 (보여지는)페이지의 시작 번호
@@ -13,6 +13,20 @@ public class Paging {
     private int tempEndPage;
 
     ///////////////////////////////////////////////////////////////////////////////////////////////
+
+    // PerPage class
+    final int PER_PAGE_NUM = 10;
+
+    private int page; // 보여줄 페이지
+    private int perPageNum; // 페이지당 보여줄  페이지수
+
+    public Paging() {
+        this.page = 1;
+        this.perPageNum = PER_PAGE_NUM;
+    }
+
+
+
     public void setTotalCount(int totalCount) {
         this.totalCount = totalCount;
 
@@ -21,11 +35,11 @@ public class Paging {
 
     private void calcData() { // 전체 필드 변수 값들을 계산하는 메서드
 
-        endPage = (int) (Math.ceil(perPage.getPage() / (double) displayPageNum) * displayPageNum);
+        endPage = (int) (Math.ceil(getPage() / (double) displayPageNum) * displayPageNum);
 
         startPage = (endPage - displayPageNum) + 1;
 
-        int tempEndPage = (int) (Math.ceil(totalCount / (double) perPage.getPerPageNum()));
+        int tempEndPage = (int) (Math.ceil(totalCount / (double) getPerPageNum()));
         this.tempEndPage = tempEndPage;
 
         if (endPage > tempEndPage) {
@@ -33,22 +47,14 @@ public class Paging {
         }
 
         prev = startPage == 1 ? false : true; // 1페이지면 이전 누를 수 없게 false
-        next = endPage * perPage.getPerPageNum() >= totalCount ? false : true;
+        next = endPage * getPerPageNum() >= totalCount ? false : true;
 
     }
 
     // getter setter
 
-    public PerPage getPerPage() {
-        return perPage;
-    }
-
     public int getTempEndPage() {
         return tempEndPage;
-    }
-
-    public void setPerPage(PerPage perPage) {
-        this.perPage = perPage;
     }
 
     public int getTotalCount() {
@@ -94,5 +100,56 @@ public class Paging {
     public void setDisplayPageNum(int displayPageNum) {
         this.displayPageNum = displayPageNum;
     }
+
+
+    //////////////////////////////////////////////////////////////////////
+    // PerPage class
+
+    // Sql 구문에서 limit 사용해야 하는데 시작부분에 필요한 값을 반환.
+    public int getPageStart() {
+        return (this.page -1) * this.perPageNum;
+    }
+
+    // 페이지당 10개!
+//    public PerPage() {
+//        this.page = 1;
+//        this.perPageNum=PER_PAGE_NUM;
+//    }
+    public void basicPage() {
+        this.page =1;
+        this.perPageNum = PER_PAGE_NUM;
+    }
+
+    public int getPage() {
+        return this.page;
+    }
+
+    public void setPage(int page) {
+        if(page < 0) {
+            this.page=1;
+        }else {
+            this.page = page;
+        }
+    }
+
+    public int getPerPageNum() {
+        return this.perPageNum;
+    }
+
+    // 왜 100보다 크다는 하는거지?
+    // 1~10 페이지에 들어가는 게시글의 수가 100개가 최대라 그런가??
+    public void setPerPageNum(int perPageNum) {
+        if(perPageNum <= 0 || perPageNum > PER_PAGE_NUM*10) {
+            this.perPageNum = PER_PAGE_NUM;
+        }else {
+            this.perPageNum = perPageNum;
+        }
+    }
+
+    @Override
+    public String toString() {
+        return "Pagin 클래스 [page=" + page + ", perPageNum=" + perPageNum + "]";
+    }
+
 
 }
