+++ src/main/java/com/fastcampus/jcjboard/servlet/ArticleListServlet.java
@@ -2,7 +2,6 @@ package com.fastcampus.jcjboard.servlet;
 
 import com.fastcampus.jcjboard.dao.BoardDao;
 import com.fastcampus.jcjboard.paging.Paging;
-import com.fastcampus.jcjboard.paging.PerPage;
 
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
@@ -17,29 +16,30 @@ import java.util.List;
 public class ArticleListServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        //int id = Integer.parseInt(req.getParameter("page"));
 
-
-        // 그냥 id 값 초기화 해준거.
         int id=1;
+        try {
+            id = Integer.parseInt(req.getParameter("page"));
+        } catch (NumberFormatException e) {
+            //id가 정수가 아닌경우 목록으로 리다이렉트
+            resp.sendRedirect("/board/list?page=1");
+            return;
+        }
 
         // 최초에 index.html 에서 넘어오면 page 값이 null 이다
         // 이후에 paging 부분의 [2].[3]을 누르면 그 인덱스가 id 로 저장이 되게 만들었어!
-        if(req.getParameter("page") != null) {
-            id = Integer.parseInt(req.getParameter("page"));
-        }
+//        if(req.getParameter("page") != null) {
+//            id = Integer.parseInt(req.getParameter("page"));
+//        }
 
         BoardDao dao = new BoardDao();
 
         // SQL limit 써서. 한 페이지당 게시글 출력 개수 조정하기!!!!!!!!!!
-        PerPage perPage = new PerPage(); // 기본값 1, 10
-        // 1 은 어차피 기본값이다.
-        if(id > 1) {
-            perPage.setPage(id);
-        }
-
         Paging paging = new Paging();
-        paging.setPerPage(perPage);
+
+        // PerPage 클래스 생성자를 대신해서.
+        paging.setPage(id);
+
         paging.setTotalCount(dao.getArticleListTotalCount());
