+++ src/main/resources/DbConf
@@ -1,4 +1,4 @@
 dbUrl = jdbc:mysql://localhost:3306/jcjboard?serverTimezone=Asia/Seoul&useSSL=false
 dbId = root
-dbPassword = 0653
+dbPassword = 1234
 dbDriver = com.mysql.cj.jdbc.Driver
